refactor(dqm_ingest): moved xref functions to ref class
